--- libimagentrymarkup/src/asciidoc/check.rs
@@ -1,6 +1,0 @@
-use libimagstore::store::Entry;
-
-pub fn is_asciidoc(e: &Entry) -> bool {
-    false
-}
-
