--- libimagentrymarkup/src/bbcode/check.rs
@@ -1,6 +1,0 @@
-use libimagstore::store::Entry;
-
-pub fn is_bbcode(e: &Entry) -> bool {
-    false
-}
-
