--- libimagentrymarkup/src/commonmark/check.rs
@@ -1,6 +1,0 @@
-use libimagstore::store::Entry;
-
-pub fn is_commonmark(e: &Entry) -> bool {
-    false
-}
-
