--- libimagentrymarkup/src/restructuredtext/check.rs
@@ -1,6 +1,0 @@
-use libimagstore::store::Entry;
-
-pub fn is_restructuredtext(e: &Entry) -> bool {
-    false
-}
-
